+++ Demo/Checkpoint.build.ps1
@@ -122,7 +122,7 @@ task TestSerialization {
 	Remove-Item z.clixml
 }
 
-# This is a test. Call the task `break` as protected because it fails. The
-# checkpoint should be created, as a result. Then call the task `resume` and
-# check that the checkpoint has been deleted.
-task test TestSerialization, @{break=1}, {assert (Test-Path checkpoint.clixml)}, resume, {assert !(Test-Path checkpoint.clixml)}
+# This is a test. Call the task `break` safe because it fails. The checkpoint
+# should be created, as a result. Then call the task `resume` and check that
+# the checkpoint has been deleted.
+task test TestSerialization, (job break -Safe), {assert (Test-Path checkpoint.clixml)}, resume, {assert !(Test-Path checkpoint.clixml)}