+++ Demo/Safe.test.ps1
@@ -1,10 +1,7 @@
 
 <#
 .Synopsis
-	Example of protected task jobs (@{Task=1} notation).
-
-.Example
-	Invoke-Build * Protected.test.ps1
+	Test safe jobs.
 #>
 
 . .\Shared.ps1
@@ -25,10 +22,10 @@ task Error2 {
 	throw "Error2"
 }
 
-# This task has two protected references to failing tasks.
+# This task has two safe references to failing tasks.
 task Survives1 @(
 	# Tells to call the task Error1 and ignore its errors
-	@{Error1=1}
+	(job Error1 -Safe)
 	# Code invoked after the task Error1
 	{
 		"After Error1"
@@ -42,7 +39,7 @@ task Survives1 @(
 		assert ($null -eq $error2)
 	}
 	# Tells to call the task Error2 and ignore its errors
-	@{Error2=1}
+	(job Error2 -Safe)
 	# Code invoked after the task Error2
 	{
 		"After Error2"
@@ -56,7 +53,7 @@ task Survives1 @(
 # Similar task. It checks that failed tasks are not called again.
 task Survives2 @(
 	# tells to call the task Error1 and ignore its failure
-	@{Error1=1}
+	(job Error1 -Safe)
 	# code invoked after the task Error1
 	{
 		"After Error1"
@@ -66,7 +63,7 @@ task Survives2 @(
 		assert ("$error1" -eq "Error1")
 	}
 	# tells to call the task Error2 and ignore its failure
-	@{Error2=1}
+	(job Error2 -Safe)
 	# code invoked after the task Error2
 	{
 		"After Error2"
@@ -79,7 +76,7 @@ task Survives2 @(
 
 ### Survives3, Survives4 use the same job list (there was an issue)
 
-$JobList = @{Error1=1}, @{Error2=1}
+$JobList = (job Error1 -Safe), (job Error2 -Safe)
 task Survives3 $JobList
 task Survives4 $JobList
 
@@ -94,27 +91,26 @@ task Error4 {throw 'Error4'}
 # it fails because the error in Error4 is going to break AlmostSurvives2.
 task AlmostSurvives1 @(
 	# Tells to call the task Error3 and ignore its failure
-	@{Error3=1},
+	(job Error3 -Safe)
 	# Code invoked after the task Error3
 	{
 		"After Error3 -- this works"
-	},
+	}
 	# Tells to call the task Error4 and ignore its failure
-	@{Error4=1},
+	(job Error4 -Safe)
 	# This code is not going to be invoked
 	{
 		throw "After Error4 -- this is not called"
 	}
 )
 
-# This task is prepared for errors in Error3, that is why build continues after
-# the first error in AlmostSurvives1. But it is not prepared for errors in
-# Error4, that is why the whole build fails. Note: it does not matter that the
-# downstream task calls this as @{AlmostSurvives2=1}, AlmostSurvives2 is not
-# ready for errors in Error4 (otherwise it would call it as @{Error4=1}).
+# This task is ready for errors in Error3, that is why build continues after
+# the first error in AlmostSurvives1. But it is not ready for errors in Error4,
+# that is why the whole build fails even though AlmostSurvives calls this task
+# safe.
 task AlmostSurvives2 @(
-	@{Error3=1},
-	{},
+	(job Error3 -Safe)
+	{}
 	# This unprotected reference makes the build to fail.
 	# IMPORTANT: This task AlmostSurvives2 is not even get called.
 	'Error4'
@@ -122,13 +118,13 @@ task AlmostSurvives2 @(
 )
 
 # This task calls the tests and fails due to issues in the AlmostSurvives2.
-# Even protected call does not help: AlmostSurvives2 is not protected from
-# errors in Error4.
-task AlmostSurvives AlmostSurvives1, @{AlmostSurvives2=1}
+# Even safe call does not help: AlmostSurvives2 is not ready for errors in
+# Error4.
+task AlmostSurvives AlmostSurvives1, (job AlmostSurvives2 -Safe)
 
 # Trigger tasks and check for expected results.
-task TestAlmostSurvives @{AlmostSurvives=1}, {
-	Test-Error AlmostSurvives "Error4*At *\Protected.test.ps1*'Error4'*OperationStopped*"
+task TestAlmostSurvives (job AlmostSurvives -Safe), {
+	Test-Error AlmostSurvives "Error4*At *\Safe.test.ps1*'Error4'*OperationStopped*"
 }
 
 ### DependsOnFailedDirectlyAndIndirectly
@@ -140,18 +136,18 @@ task FailedUsedByMany {
 task DependsOnFailed FailedUsedByMany, {
 	throw 'Must not be called'
 }
-task DependsOnFailedDirectlyAndIndirectly @{FailedUsedByMany=1}, @{DependsOnFailed=1}, {
+task DependsOnFailedDirectlyAndIndirectly (job FailedUsedByMany -Safe), (job DependsOnFailed -Safe), {
 	throw 'Must not be called'
 }
-task TestDependsOnFailedDirectlyAndIndirectly @{DependsOnFailedDirectlyAndIndirectly=1}, {
+task TestDependsOnFailedDirectlyAndIndirectly (job DependsOnFailedDirectlyAndIndirectly -Safe), {
 	# error of initial failure
 	assert ("$(error FailedUsedByMany)" -eq 'Oops in FailedUsedByMany')
 
-	# no error because it is not called, even if it is called protected itself
-	# it also calls the failed task not protected
+	# no error because it is not called, even if it is called safe itself it
+	# also calls the failed task unsafe
 	assert !(error DependsOnFailed)
 
-	# error, even if it calls the failed task protected it also calls another
-	# task which leads to not protected calls of the failed task
+	# error, even if it calls the failed task safe it also calls another task
+	# which leads to unsafe calls of the failed task
 	assert ("$(error DependsOnFailedDirectlyAndIndirectly)" -eq 'Oops in FailedUsedByMany')
 }